+++ ext/mbstring/libmbfl/filters/mbfilter_utf8.h
@@ -34,8 +34,18 @@ extern const mbfl_encoding mbfl_encoding_utf8;
 extern const struct mbfl_convert_vtbl vtbl_utf8_wchar;
 extern const struct mbfl_convert_vtbl vtbl_wchar_utf8;
 
-int mbfl_filt_conv_utf8_wchar(int c, mbfl_convert_filter *filter);
-int mbfl_filt_conv_wchar_utf8(int c, mbfl_convert_filter *filter);
-int mbfl_filt_conv_utf8_wchar_flush(mbfl_convert_filter *filter);
+extern const mbfl_encoding mbfl_encoding_utf8_docomo;
+extern const mbfl_encoding mbfl_encoding_utf8_kddi_a;
+extern const mbfl_encoding mbfl_encoding_utf8_kddi_b;
+extern const mbfl_encoding mbfl_encoding_utf8_sb;
+
+extern const struct mbfl_convert_vtbl vtbl_utf8_docomo_wchar;
+extern const struct mbfl_convert_vtbl vtbl_wchar_utf8_docomo;
+extern const struct mbfl_convert_vtbl vtbl_utf8_kddi_a_wchar;
+extern const struct mbfl_convert_vtbl vtbl_wchar_utf8_kddi_a;
+extern const struct mbfl_convert_vtbl vtbl_utf8_kddi_b_wchar;
+extern const struct mbfl_convert_vtbl vtbl_wchar_utf8_kddi_b;
+extern const struct mbfl_convert_vtbl vtbl_utf8_sb_wchar;
+extern const struct mbfl_convert_vtbl vtbl_wchar_utf8_sb;
 
 #endif /* MBFL_MBFILTER_UTF8_H */
